[기능 구현] 김진도 - 학생 퀴즈방 입장 , 퀴즈 문제 고르기

+++ src/main/java/com/samsamgyeesam/studyingvally/domain/quiz/dto/QuizChapterDTO.java
@@ -8,9 +8,9 @@ import lombok.*;
 @Setter
 @ToString
 public class QuizChapterDTO {
-    private int chapNo;
+    private Long chapNo;
     private String chapTitle;
     private String chapDesc;
     private String chapUrl;
-    private Integer courseId;
+    private Long courseId;
 }
